+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [15.9.0] (melonJS 2) - _2023-08-xx_
 
 ### Added
+- Input: new `hasActiveEvents` returning true if there are any pending events in the queue
 - Renderer: fix/enable batching for all ellipse & arc(To) WebGL stroke operations
 
 ## [15.8.0] (melonJS 2) - _2023-07-29_